apps/audio-analyzer/requirements.txt
# Audio processing
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.11.0

# API server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# DTW alignment
dtw-python>=1.3.0

# Audio file handling
soundfile>=0.12.0
pydub>=0.25.0

# HTTP client for fetching reference audio
httpx>=0.26.0
aiofiles>=23.2.0
